backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
watchdog==4.0.0
GitPython==3.1.41
chromadb==0.4.22
numpy<2.0.0
psutil==5.9.8
pytest==8.0.0
httpx==0.26.0